+++ packages/components/icon/src/icon.tsx
@@ -1,10 +1,10 @@
 import type {
   CSSUIObject,
   CSSUIProps,
+  FC,
   HTMLUIProps,
   ThemeProps,
 } from "@yamada-ui/core"
-import type { FC } from "react"
 import {
   forwardRef,
   mergeVars,
@@ -79,6 +79,9 @@ export const Icon = forwardRef<IconProps, "svg">((props, ref) => {
   )
 })
 
+Icon.displayName = "Icon"
+Icon.__ui__ = "Icon"
+
 export const CheckIcon: FC<IconProps> = (props) => {
   return (
     <Icon viewBox="0 0 24 24" {...props}>
@@ -89,6 +92,9 @@ export const CheckIcon: FC<IconProps> = (props) => {
     </Icon>
   )
 }
+
+CheckIcon.displayName = "CheckIcon"
+CheckIcon.__ui__ = "CheckIcon"
 
 export const InfoIcon: FC<IconProps> = (props) => {
   return (
@@ -101,6 +107,9 @@ export const InfoIcon: FC<IconProps> = (props) => {
   )
 }
 
+InfoIcon.displayName = "InfoIcon"
+InfoIcon.__ui__ = "InfoIcon"
+
 export const WarningIcon: FC<IconProps> = (props) => {
   return (
     <Icon viewBox="0 0 24 24" {...props}>
@@ -111,6 +120,9 @@ export const WarningIcon: FC<IconProps> = (props) => {
     </Icon>
   )
 }
+
+WarningIcon.displayName = "WarningIcon"
+WarningIcon.__ui__ = "WarningIcon"
 
 export const CloseIcon: FC<IconProps> = (props) => {
   return (
@@ -123,6 +135,9 @@ export const CloseIcon: FC<IconProps> = (props) => {
   )
 }
 
+CloseIcon.displayName = "CloseIcon"
+CloseIcon.__ui__ = "CloseIcon"
+
 export const ChevronIcon: FC<IconProps> = (props) => {
   return (
     <Icon aria-hidden focusable="false" viewBox="0 0 24 24" {...props}>
@@ -133,3 +148,6 @@ export const ChevronIcon: FC<IconProps> = (props) => {
     </Icon>
   )
 }
+
+ChevronIcon.displayName = "ChevronIcon"
+ChevronIcon.__ui__ = "ChevronIcon"