+++ packages/components/table/src/use-table.tsx
@@ -1,7 +1,9 @@
 import type {
   AccessorColumnDef,
   Cell,
+  CellContext,
   ColumnDef,
+  ColumnDefTemplate,
   CoreOptions,
   DisplayColumnDef,
   Header,
@@ -12,7 +14,6 @@ import type {
   Row,
   RowData,
   RowSelectionOptions,
-  RowSelectionState,
   SortingOptions,
   SortingState,
 } from "@tanstack/react-table"
@@ -25,24 +26,29 @@ import type {
 } from "@yamada-ui/core"
 import type { IconProps } from "@yamada-ui/icon"
 import type { TdProps, ThProps, TrProps } from "@yamada-ui/native-table"
-import type { CSSProperties } from "react"
+import type {
+  CSSProperties,
+  KeyboardEvent,
+  KeyboardEventHandler,
+  RefObject,
+} from "react"
+import type { SelectColumn } from "./table-utils"
 import {
   getCoreRowModel,
   getPaginationRowModel,
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table"
-import { Checkbox } from "@yamada-ui/checkbox"
-import { ui } from "@yamada-ui/core"
 import { useControllableState } from "@yamada-ui/use-controllable-state"
+import { runIfFunc } from "@yamada-ui/utils"
+import { useCallback, useMemo, useRef, useState } from "react"
 import {
-  ariaAttr,
-  createContext,
-  handlerAll,
-  runIfFunc,
-} from "@yamada-ui/utils"
-import { useCallback, useMemo } from "react"
-
+  generateGroups,
+  generateRowSelection,
+  getEnableRowSelection,
+  mergeColumns,
+} from "./table-utils"
+import { generateRowId } from "./table-utils"
 export { createColumnHelper, flexRender as render } from "@tanstack/react-table"
 export type {
   Cell,
@@ -52,12 +58,7 @@ export type {
   SortDirection,
 } from "@tanstack/react-table"
 
-export type TableContext = Omit<UseTableReturn, "getTableProps">
-
-export const [TableProvider, useTableContext] = createContext<TableContext>({
-  name: "TableContext",
-  errorMessage: `useTableContext returned is 'undefined'. Seems you forgot to wrap the components in "<Table />"`,
-})
+export type CellMap = Map<string, RefObject<HTMLTableCellElement>>
 
 export interface PropsColumnDef {
   className?: string
@@ -73,19 +74,31 @@ export interface GroupColumnDef<Y extends RowData, M = any> {
   columns?: Column<Y, M>[]
 }
 
+interface ColumnProps {
+  referenceRef: RefObject<any>
+  tabIndex: number
+}
+
+interface ColumnBase<Y extends RowData, M = any> {
+  cell?: ColumnDefTemplate<CellContext<Y, M> & ColumnProps>
+  footer?: ColumnDefTemplate<ColumnProps & HeaderContext<Y, M>>
+  header?: ColumnDefTemplate<ColumnProps & HeaderContext<Y, M>>
+}
+
 export type Column<Y extends RowData, M = any> = (
+  | Omit<AccessorColumnDef<Y, M>, keyof ColumnBase<Y, M>>
+  | Omit<DisplayColumnDef<Y, M>, keyof ColumnBase<Y, M>>
+) &
+  ColumnBase<Y, M> &
+  GroupColumnDef<Y, M> &
+  PropsColumnDef
+
+export type DefaultColumn<Y extends RowData, M = any> = (
   | AccessorColumnDef<Y, M>
   | DisplayColumnDef<Y, M>
 ) &
   GroupColumnDef<Y, M> &
   PropsColumnDef
-
-export type InternalColumn<Y extends RowData, M = any> = {
-  "aria-hidden"?: boolean
-} & Column<Y, M>
-
-interface SelectColumn<Y extends RowData, M = any>
-  extends Omit<Column<Y, M>, "accessorFn" | "accessorKey"> {}
 
 export interface ColumnSort<Y extends RowData> {
   id: keyof Y
@@ -145,7 +158,7 @@ export interface UseTableProps<Y extends RowData>
   /**
    * Default column options to use for all column defs supplied to the table.
    */
-  defaultColumn?: Partial<Column<Y>>
+  defaultColumn?: Partial<DefaultColumn<Y>>
   /**
    * The initial page index of the paging table.
    *
@@ -177,6 +190,12 @@ export interface UseTableProps<Y extends RowData>
    */
   enablePagination?: boolean
   /**
+   * If `true`, the table cell will be focusable.
+   *
+   * @default true
+   */
+  isFocusable?: boolean
+  /**
    * The page index of the paging table.
    */
   pageIndex?: number
@@ -190,6 +209,11 @@ export interface UseTableProps<Y extends RowData>
    * @default '[20, 50, 100]'
    */
   pageSizeList?: number[]
+  /**
+   * The role used for the row header.
+   * This is used for accessibility to announce the selected row.
+   */
+  rowHeader?: keyof Y
   /**
    * The id used to store the value when selected.
    */
@@ -208,6 +232,12 @@ export interface UseTableProps<Y extends RowData>
    * The sort of the table.
    */
   sort?: Sort<Y>
+  /**
+   * If `true`, display the table footer.
+   *
+   * @default false
+   */
+  withFooter?: boolean
   /**
    * If `true`, display the checkbox in table footer.
    *
@@ -276,30 +306,6 @@ export interface UseTableProps<Y extends RowData>
   onDoubleClickRow?: (row: Row<Y>) => void
 }
 
-const generateRowSelection = <Y extends RowData>(
-  rowSelection: string[] | undefined,
-  enableRowSelection: UseTableProps<Y>["enableRowSelection"],
-): RowSelectionState => {
-  if (!enableRowSelection) return {}
-
-  if (rowSelection) {
-    return rowSelection.reduce<RowSelectionState>(
-      (prev, id) => ({ ...prev, [String(id)]: true }),
-      {},
-    )
-  } else {
-    return {}
-  }
-}
-
-const generateRowId = <Y extends RowData>(key: keyof Y | undefined) =>
-  key ? (row: Y) => String(row[key]) : undefined
-
-const computedEnableRowSelection = <Y extends RowData>(
-  { id }: Row<Y>,
-  disabledRowIds?: string[],
-) => !disabledRowIds?.includes(id)
-
 export const useTable = <Y extends RowData>({
   autoResetAll,
   autoResetPageIndex,
@@ -319,11 +325,12 @@ export const useTable = <Y extends RowData>({
   enableMultiRemove,
   enableMultiSort,
   enablePagination = false,
-  enableRowSelection = (row) => computedEnableRowSelection(row, disabledRowIds),
+  enableRowSelection = (row) => getEnableRowSelection(row, disabledRowIds),
   enableSorting,
   enableSortingRemoval,
   rowId,
   getRowId = generateRowId(rowId),
+  isFocusable = true,
   isMultiSortEvent,
   manualPagination,
   manualSorting,
@@ -334,11 +341,13 @@ export const useTable = <Y extends RowData>({
   pageSize,
   pageSizeList = [20, 50, 100],
   renderFallbackValue,
+  rowHeader,
   rowsClickSelect,
   selectedRowIds,
   sort,
   sortDescFirst,
   sortingFns,
+  withFooter = false,
   withFooterSelect,
   cellProps,
   checkboxProps,
@@ -357,24 +366,23 @@ export const useTable = <Y extends RowData>({
   onDoubleClickRow,
   ...rest
 }: UseTableProps<Y>) => {
+  const cellMapRef = useRef<CellMap>(new Map())
+  const [focusedCell, setFocusedCell] = useState<[number, number]>([0, 0])
   const [sorting, onSortingChange] = useControllableState({
     defaultValue: defaultSort,
     value: sort,
     onChange: onChangeSort,
   }) as unknown as [SortingState, OnChangeFn<SortingState>]
-
   const [rowSelection, onRowSelectionChange] = useControllableState({
     defaultValue: defaultSelectedRowIds,
     value: selectedRowIds,
     onChange: onChangeSelect,
   })
-
   const [internalPageIndex, setInternalPageIndex] = useControllableState({
     defaultValue: defaultPageIndex,
     value: pageIndex,
     onChange: onChangePageIndex,
   })
-
   const [internalPageSize, setInternalPageSize] = useControllableState({
     defaultValue: defaultPageSize,
     value: pageSize,
@@ -406,12 +414,14 @@ export const useTable = <Y extends RowData>({
             columns,
             disabledRowIds,
             enablePagination,
+            rowHeader,
             withFooterSelect,
             checkboxProps,
             selectColumnProps,
           })
         : columns,
     [
+      rowHeader,
       checkboxProps,
       columns,
       disabledRowIds,
@@ -428,6 +438,7 @@ export const useTable = <Y extends RowData>({
   )
 
   const {
+    getAllFlatColumns,
     getCanNextPage,
     getCanPreviousPage,
     getFooterGroups,
@@ -505,21 +516,109 @@ export const useTable = <Y extends RowData>({
   )
 
   const state = getState()
-  const headerGroups = getHeaderGroups()
-  const footerGroups = getFooterGroups()
+  const _headerGroups = getHeaderGroups()
+  const _footerGroups = getFooterGroups()
+  const flatColumns = getAllFlatColumns()
   const { flatRows, rows, rowsById } = getRowModel()
   const totalPage = getPageCount()
 
+  const headerGroups = useMemo(
+    () => generateGroups<Y>(_headerGroups),
+    [_headerGroups],
+  )
+
+  const footerGroups = useMemo(
+    () => generateGroups<Y>(_footerGroups, true),
+    [_footerGroups],
+  )
+
+  const headerCount = headerGroups.length
+  const rowCount = rows.length
+  const footerCount = withFooter ? footerGroups.length : 0
+  const colCount = flatColumns.length
+  const maxRowCount = headerCount + rowCount + footerCount - 1
+  const maxColIndex = colCount - 1
+
+  const getOnKeyDown = useCallback(
+    (rowIndex: number, colIndex: number) =>
+      (ev: KeyboardEvent<HTMLTableCellElement>) => {
+        const onFocusCell = (nextRowIndex: number, nextColIndex: number) => {
+          const currentEl = cellMapRef.current.get(`${rowIndex}-${colIndex}`)
+          const nextEl = cellMapRef.current.get(
+            `${nextRowIndex}-${nextColIndex}`,
+          )
+          const isNotfound = !nextEl
+          const isSome = currentEl === nextEl
+
+          if (isSome || isNotfound) {
+            const onKeyDown = getOnKeyDown(nextRowIndex, nextColIndex)
+
+            onKeyDown(ev)
+          } else {
+            setFocusedCell([nextRowIndex, nextColIndex])
+
+            setTimeout(() => {
+              cellMapRef.current
+                .get(`${nextRowIndex}-${nextColIndex}`)
+                ?.current?.focus()
+            })
+          }
+        }
+
+        const actions: { [key: string]: KeyboardEventHandler } = {
+          ArrowDown: () => {
+            if (rowIndex >= maxRowCount) return
+
+            onFocusCell(rowIndex + 1, colIndex)
+          },
+          ArrowLeft: () => {
+            if (colIndex <= 0) return
+
+            onFocusCell(rowIndex, colIndex - 1)
+          },
+          ArrowRight: () => {
+            if (colIndex >= maxColIndex) return
+
+            onFocusCell(rowIndex, colIndex + 1)
+          },
+          ArrowUp: () => {
+            if (rowIndex <= 0) return
+
+            onFocusCell(rowIndex - 1, colIndex)
+          },
+          End: () => {
+            onFocusCell(rowIndex, maxColIndex)
+          },
+          Home: () => {
+            onFocusCell(rowIndex, 0)
+          },
+        }
+
+        const action = actions[ev.key]
+
+        if (!action) return
+
+        ev.preventDefault()
+        action(ev)
+      },
+    [maxColIndex, maxRowCount],
+  )
+
   return {
+    cellMapRef,
     enableRowSelection,
     flatRows,
+    focusedCell,
     footerGroups,
     getCanNextPage,
     getCanPreviousPage,
+    getOnKeyDown,
     headerGroups,
+    isFocusable,
     nextPage,
     pageSizeList: computedPageSizeList,
     previousPage,
+    rowHeader,
     rows,
     rowsById,
     rowsClickSelect,
@@ -527,6 +626,7 @@ export const useTable = <Y extends RowData>({
     setPageSize,
     state,
     totalPage,
+    withFooter,
     cellProps,
     footerGroupProps,
     footerProps,
@@ -540,124 +640,4 @@ export const useTable = <Y extends RowData>({
   }
 }
 
-export type UseTableReturn = ReturnType<typeof useTable>
-
-const Center = ui("div", {
-  baseStyle: {
-    alignItems: "center",
-    display: "flex",
-    h: "100%",
-    justifyContent: "center",
-    w: "100%",
-  },
-})
-
-const TotalCheckbox = <Y extends RowData>({
-  disabledRowIds = [],
-  enablePagination,
-  table,
-  checkboxProps,
-}: {
-  enablePagination: boolean
-  table: HeaderContext<Y, unknown>["table"]
-  checkboxProps: CheckboxProps
-  disabledRowIds?: string[]
-}) => {
-  const {
-    getIsAllPageRowsSelected,
-    getIsAllRowsSelected,
-    getIsSomePageRowsSelected,
-    getIsSomeRowsSelected,
-    getRowModel,
-    getState,
-    getToggleAllPageRowsSelectedHandler,
-    getToggleAllRowsSelectedHandler,
-  } = table
-
-  const state = getState()
-  const { rows } = getRowModel()
-  const rowIds = rows.map(({ id }) => id)
-  const selectedRowIds = Object.keys(state.rowSelection)
-  const unselectedRowIds = rowIds.filter((id) => !selectedRowIds.includes(id))
-
-  const isAllChecked = unselectedRowIds.every((id) =>
-    disabledRowIds.includes(id),
-  )
-  const isChecked = !enablePagination
-    ? getIsAllRowsSelected()
-    : getIsAllPageRowsSelected()
-  const isIndeterminate = !enablePagination
-    ? getIsSomeRowsSelected()
-    : getIsSomePageRowsSelected()
-  const onChange = !enablePagination
-    ? getToggleAllRowsSelectedHandler()
-    : getToggleAllPageRowsSelectedHandler()
-
-  return (
-    <Center>
-      <Checkbox
-        inputProps={{ "aria-label": "Select all row" }}
-        {...{ gap: 0, ...checkboxProps }}
-        isChecked={isAllChecked || isChecked}
-        {...(!isAllChecked ? { isIndeterminate } : {})}
-        onChange={handlerAll(checkboxProps.onChange, onChange)}
-      />
-    </Center>
-  )
-}
-
-export const mergeColumns = <Y extends RowData>({
-  columns,
-  disabledRowIds,
-  enablePagination,
-  withFooterSelect,
-  checkboxProps = {},
-  selectColumnProps,
-}: {
-  columns: Column<Y>[]
-  enablePagination: boolean
-  disabledRowIds?: string[]
-  withFooterSelect?: boolean
-  checkboxProps?: CheckboxProps
-  selectColumnProps?: SelectColumn<Y>
-}): InternalColumn<Y>[] => [
-  {
-    id: "select",
-    header: ({ table }) => (
-      <TotalCheckbox
-        {...{ disabledRowIds, enablePagination, table, checkboxProps }}
-      />
-    ),
-    ...(withFooterSelect
-      ? {
-          footer: ({ table }) => (
-            <TotalCheckbox
-              {...{ disabledRowIds, enablePagination, table, checkboxProps }}
-            />
-          ),
-        }
-      : {}),
-    "aria-hidden": ariaAttr(!withFooterSelect),
-    cell: ({ row }) => {
-      const { getCanSelect, getIsSelected, getToggleSelectedHandler } = row
-
-      return (
-        <Center>
-          <Checkbox
-            inputProps={{ "aria-label": "Select row" }}
-            {...{ gap: 0, ...checkboxProps }}
-            isChecked={getIsSelected()}
-            isDisabled={!getCanSelect()}
-            onChange={handlerAll(
-              checkboxProps.onChange,
-              getToggleSelectedHandler(),
-            )}
-          />
-        </Center>
-      )
-    },
-    ...selectColumnProps,
-    css: { w: "0", ...selectColumnProps?.css },
-  },
-  ...columns,
-]
+export type UseTableReturn<Y extends RowData> = ReturnType<typeof useTable<Y>>