+++ packages/components/file-input/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @yamada-ui/file-input
 
+## 1.0.44
+
+### Patch Changes
+
+- [#3033](https://github.com/yamada-ui/yamada-ui/pull/3033) [`afe1bdc`](https://github.com/yamada-ui/yamada-ui/commit/afe1bdcbd865a1bea1ff457d6f7c50628f7d45e2) Thanks [@cidkumagai](https://github.com/cidkumagai)! - Set `displayName` and `__ui__` for `@yamada-ui/file-input`
+
+- Updated dependencies [[`898ab83`](https://github.com/yamada-ui/yamada-ui/commit/898ab83eb677646ee293a336a0a5d254aeefa00f), [`2173466`](https://github.com/yamada-ui/yamada-ui/commit/2173466c144212522a61aec94e34cac375291655), [`28632a8`](https://github.com/yamada-ui/yamada-ui/commit/28632a8d904b581068e93a75e44323c9d37111b1), [`c22d335`](https://github.com/yamada-ui/yamada-ui/commit/c22d3355ab42d4f0f3e3413248bed949729fc8e3), [`2173466`](https://github.com/yamada-ui/yamada-ui/commit/2173466c144212522a61aec94e34cac375291655)]:
+  - @yamada-ui/core@1.15.2
+  - @yamada-ui/utils@1.5.3
+  - @yamada-ui/use-controllable-state@1.0.22
+  - @yamada-ui/form-control@2.1.4
+
 ## 1.0.43
 
 ### Patch Changes