+++ packages/components/segmented-control/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @yamada-ui/segmented-control
 
+## 1.0.42
+
+### Patch Changes
+
+- [#3020](https://github.com/yamada-ui/yamada-ui/pull/3020) [`8600b39`](https://github.com/yamada-ui/yamada-ui/commit/8600b39bb6cc164bb1860f9708cd2613d49d13d5) Thanks [@0xtkey256](https://github.com/0xtkey256)! - Set `displayName` and `__ui__` after each component definition.
+
+- Updated dependencies [[`898ab83`](https://github.com/yamada-ui/yamada-ui/commit/898ab83eb677646ee293a336a0a5d254aeefa00f), [`2173466`](https://github.com/yamada-ui/yamada-ui/commit/2173466c144212522a61aec94e34cac375291655), [`28632a8`](https://github.com/yamada-ui/yamada-ui/commit/28632a8d904b581068e93a75e44323c9d37111b1), [`c22d335`](https://github.com/yamada-ui/yamada-ui/commit/c22d3355ab42d4f0f3e3413248bed949729fc8e3), [`2173466`](https://github.com/yamada-ui/yamada-ui/commit/2173466c144212522a61aec94e34cac375291655)]:
+  - @yamada-ui/core@1.15.2
+  - @yamada-ui/utils@1.5.3
+  - @yamada-ui/use-controllable-state@1.0.22
+  - @yamada-ui/motion@2.2.4
+  - @yamada-ui/use-descendant@1.0.23
+  - @yamada-ui/use-focus-visible@1.1.9
+
 ## 1.0.41
 
 ### Patch Changes