+++ packages/hooks/use-infinite-scroll/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @yamada-ui/use-infinite-scroll
 
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [[`2173466`](https://github.com/yamada-ui/yamada-ui/commit/2173466c144212522a61aec94e34cac375291655)]:
+  - @yamada-ui/utils@1.5.3
+
 ## 1.1.5
 
 ### Patch Changes